fix: django waits for my mysql to initalize
Currently the django doesn't wait for the mysql to initialize. Since the first execution of the mysql is slower (it loads the data), the django fails to initialize, which required to execute it again. 
Now the django entrypoint checks the state of the mysql database periodically and just initializes if the database is initialized.